Create a prompt UI
