+++ iterate/src/main/res/layout/fragment_prompt.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingBottom="20dp"
-    tools:context="com.iteratehq.iterate.Prompt">
+    tools:context="com.iteratehq.iterate.PromptFragment">
 
     <ImageButton
         android:id="@+id/btn_close"
@@ -27,10 +27,10 @@
         android:layout_marginRight="40dp"
         android:gravity="center_horizontal"
         android:lineSpacingExtra="4sp"
-        android:text="Help us improve the Iterate iOS SDK by answering a few questions about your ideal use cases"
         android:textColor="@android:color/black"
         android:textSize="16sp"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toTopOf="parent"
+        tools:text="Help us improve the Iterate iOS SDK by answering a few questions about your ideal use cases" />
 
     <Button
         android:id="@+id/btn_prompt"
@@ -40,12 +40,10 @@
         android:layout_marginTop="16dp"
         android:layout_marginRight="40dp"
         android:backgroundTint="#7457BE"
-        android:insetTop="0dp"
-        android:insetBottom="0dp"
-        android:text="Happy to Help"
         android:textAllCaps="false"
         android:textColor="@android:color/white"
         app:cornerRadius="6dp"
-        app:layout_constraintTop_toBottomOf="@id/txt_prompt" />
+        app:layout_constraintTop_toBottomOf="@id/txt_prompt"
+        tools:text="Happy to Help" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
